// frontend/app/imaging/viewer/shared/types.ts
export interface Tool {
  id: string;
  name: string;
  icon: string;
  description: string;
  pointsNeeded: number;
}